backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
httpx==0.27.2
celery==5.4.0
redis==5.1.1
librosa==0.10.2
numpy==1.26.4
scipy==1.14.1
scikit-learn==1.5.2
reportlab==4.2.2
matplotlib==3.9.2
python-multipart==0.0.12
aiofiles==24.1.0
python-dotenv==1.0.1
sse-starlette==2.1.3
Pillow==10.4.0
soundfile==0.12.1
